DEV-31 tanfolyamkereso kozos header refactor
