+++ components/TanfolyamkeresoCard.vue
